Add callback to adjust updateConfigurations
using a .htaccess file which blocks everything in the update directory
expect the "updates.json", it's possible to write a php script which
returns the update package file as the only possility to update the
software. A license check can be added for example. Now, the update uris
must only be changed to the php file which is possible through this
callback and everything is awesome.